refactor: vendor packagehallucination helpers per-tool
Address #1786 review feedback: tools/packagehallucination/* are
intentionally standalone scripts that ship outside the python package
and run in minimal environments. Replace the shared _common.py with
inlined copies of the helpers (normalise_first_seen, emit_record,
configure_argparse, write_jsonl) in each tool's main.py, so each tool
keeps its self-contained shape while still emitting the same standard
{text, package_first_seen} JSONL schema and exposing the same
--output / --format / --help CLI surface called out by #1745.

Delete _common.py and the shared test scaffolding (the tools are not
imported as a package; the inlined helpers are simple enough that the
per-tool standalone shape is its own contract).

Verified that all three tools produce identical record schema and
identical CLI flag set with the inlined helpers.
